plugins/memory-session-manager/requirements.txt
mem0ai==1.0.1
qdrant-client==1.16.0
